Remove dead / legacy code - part 1
