--- operator/pkg/registry/labels.go
@@ -1,12 +1,0 @@
-package registry
-
-func applyRegistryLabels(labels map[string]string, component string) map[string]string {
-	if labels == nil {
-		labels = make(map[string]string)
-	}
-	labels["app"] = "docker-registry" // this is the backup/restore label for the registry
-	labels["app.kubernetes.io/component"] = component
-	labels["app.kubernetes.io/part-of"] = "embedded-cluster"
-	labels["app.kubernetes.io/managed-by"] = "embedded-cluster-operator"
-	return labels
-}
