--- pkg/addons/adminconsole/static/metadata.yaml
@@ -1,30 +1,0 @@
-#
-# this file is automatically generated by buildtools. manual edits are not recommended.
-# to regenerate this file, run the following commands:
-#
-# $ make buildtools
-# $ output/bin/buildtools update addon <addon name>
-#
-version: 1.124.1
-location: oci://proxy.replicated.com/anonymous/registry.replicated.com/library/admin-console
-images:
-    kotsadm:
-        repo: proxy.replicated.com/anonymous/kotsadm/kotsadm
-        tag:
-            amd64: v1.124.1-amd64@sha256:12c31ac4a7e56deba51a385852043dde620fa19ce4a2b538fa46f495d377807b
-            arm64: v1.124.1-arm64@sha256:71295831280ce031c06814edda7c523583a8f18f8c991fc76ce8f0b065d6afc0
-    kotsadm-migrations:
-        repo: proxy.replicated.com/anonymous/kotsadm/kotsadm-migrations
-        tag:
-            amd64: v1.124.1-amd64@sha256:093f99c63e12179eb2385a2c153526991a3092ff34b58c621724466a7c26c763
-            arm64: v1.124.1-arm64@sha256:1bccf571a3334825923eed7a2bb4482496201917f813f7b4bf3a3f33beacd9cb
-    kurl-proxy:
-        repo: proxy.replicated.com/anonymous/kotsadm/kurl-proxy
-        tag:
-            amd64: v1.124.1-amd64@sha256:ec1aae690fa363083eea2c62685b56d372c3bf8ec4a5ac8a70d387270e2027b1
-            arm64: v1.124.1-arm64@sha256:8b52daf2b132b91d5db948cc95c32fec5620af1068cfdf4756e025a327bd4541
-    rqlite:
-        repo: proxy.replicated.com/anonymous/kotsadm/rqlite
-        tag:
-            amd64: 8.36.10-r0-amd64@sha256:6eb9ab620632fc3e1697bc8619fe5f2fe3d5a6fa19b6e3f84b35cca1f6254ef3
-            arm64: 8.36.10-r0-arm64@sha256:259569537110f8c1409863fb35868925f1c3b482689d11b4225285f361c1259d
