--- pkg/addons/adminconsole/static/values.tpl.yaml
@@ -1,24 +1,0 @@
-{{- if .ReplaceImages }}
-images:
-    kotsadm: '{{ ImageString (index .Images "kotsadm") }}'
-    kurlProxy: '{{ ImageString (index .Images "kurl-proxy") }}'
-    migrations: '{{ ImageString (index .Images "kotsadm-migrations") }}'
-    rqlite: '{{ ImageString (index .Images "rqlite") }}'
-{{- end }}
-isHA: false
-isHelmManaged: false
-kurlProxy:
-    enabled: true
-    nodePort: 30000
-labels:
-    replicated.com/disaster-recovery: infra
-    replicated.com/disaster-recovery-chart: admin-console
-minimalRBAC: false
-passwordSecretRef:
-    key: passwordBcrypt
-    name: kotsadm-password
-service:
-    enabled: false
-privateCAs:
-    enabled: true
-    configmapName: "kotsadm-private-cas"
