--- pkg/addons/embeddedclusteroperator/static/metadata.yaml
@@ -1,20 +1,0 @@
-#
-# this file is automatically generated by buildtools. manual edits are not recommended.
-# to regenerate this file, run the following commands:
-#
-# $ make buildtools
-# $ output/bin/buildtools update addon <addon name>
-#
-version: 1.19.0+k8s-1.30
-location: oci://proxy.replicated.com/anonymous/registry.replicated.com/library/embedded-cluster-operator
-images:
-    embedded-cluster-operator:
-        repo: proxy.replicated.com/anonymous/replicated/embedded-cluster-operator-image
-        tag:
-            amd64: v1.19.0-k8s-1.30
-            arm64: v1.19.0-k8s-1.30
-    utils:
-        repo: proxy.replicated.com/anonymous/replicated/ec-utils
-        tag:
-            amd64: latest
-            arm64: latest
