--- pkg/addons/embeddedclusteroperator/static/values.tpl.yaml
@@ -1,10 +1,0 @@
-global:
-  labels:
-    replicated.com/disaster-recovery: infra
-    replicated.com/disaster-recovery-chart: embedded-cluster-operator
-{{- if .ReplaceImages }}
-image:
-  repository: '{{ (index .Images "embedded-cluster-operator").Repo }}'
-  tag: '{{ index (index .Images "embedded-cluster-operator").Tag .GOARCH }}'
-utilsImage: '{{ ImageString (index .Images "utils") }}'
-{{- end }}
