--- pkg/addons/openebs/static/metadata.yaml
@@ -1,25 +1,0 @@
-#
-# this file is automatically generated by buildtools. manual edits are not recommended.
-# to regenerate this file, run the following commands:
-#
-# $ make buildtools
-# $ output/bin/buildtools update addon <addon name>
-#
-version: 4.1.2
-location: oci://proxy.replicated.com/anonymous/registry.replicated.com/ec-charts/openebs
-images:
-    kubectl:
-        repo: proxy.replicated.com/anonymous/replicated/ec-kubectl
-        tag:
-            amd64: 1.32.1-r2-amd64@sha256:479452dd00b3d2b17b5e0a93ca6f376a781608e906238685376ab233a6f1af7c
-            arm64: 1.32.1-r2-arm64@sha256:df0d96453f2fe340801f1031419245f97a22d5a4abaa5b7039fc5fd894d34f64
-    openebs-linux-utils:
-        repo: proxy.replicated.com/anonymous/replicated/ec-openebs-linux-utils
-        tag:
-            amd64: 4.1.2-amd64@sha256:c2bd68b4ad8a53cc8ad71b8250e334c2e693e27fc2f9d3104e3cf69028b8562b
-            arm64: 4.1.2-arm64@sha256:a855a84d4dda5e6f17fa452fd726e3816be6685034e7c332e57092c04057cf4d
-    openebs-provisioner-localpv:
-        repo: proxy.replicated.com/anonymous/replicated/ec-openebs-provisioner-localpv
-        tag:
-            amd64: 4.1.4-r0-amd64@sha256:37ac863546f51d24966d890099a9d222b9f09a541abad26962d44d1068f11de0
-            arm64: 4.1.4-r0-arm64@sha256:1c000074fe4b288a0ec2af4c7bff34312b5d02e9b7e89c6457d2957f02496507
