--- pkg/addons/registry/static/metadata.yaml
@@ -1,15 +1,0 @@
-#
-# this file is automatically generated by buildtools. manual edits are not recommended.
-# to regenerate this file, run the following commands:
-#
-# $ make buildtools
-# $ output/bin/buildtools update addon <addon name>
-#
-version: 2.2.3
-location: oci://proxy.replicated.com/anonymous/registry.replicated.com/ec-charts/docker-registry
-images:
-    registry:
-        repo: proxy.replicated.com/anonymous/replicated/ec-registry
-        tag:
-            amd64: 2.8.3-r0@sha256:5b76ebd0a362009e31a05ac487c690f5ece0e11f6c4d9261ca63a3f162b57660
-            arm64: 2.8.3-r0-arm64@sha256:e9841478ded46f9ed3f258ea7b291718954cecce2d6907df778d223dded2dd9d
