--- pkg/addons/registry/static/values-ha.tpl.yaml
@@ -1,47 +1,0 @@
-affinity:
-  podAntiAffinity:
-    requiredDuringSchedulingIgnoredDuringExecution:
-    - labelSelector:
-        matchExpressions:
-        - key: app
-          operator: In
-          values:
-          - docker-registry
-      topologyKey: kubernetes.io/hostname
-configData:
-  auth:
-    htpasswd:
-      path: /auth/htpasswd
-      realm: Registry
-  storage:
-    s3:
-      secure: false
-extraVolumeMounts:
-- mountPath: /auth
-  name: auth
-extraVolumes:
-- name: auth
-  secret:
-    secretName: registry-auth
-fullnameOverride: registry
-{{- if .ReplaceImages }}
-image:
-  repository: '{{ (index .Images "registry").Repo }}'
-  tag: '{{ index (index .Images "registry").Tag .GOARCH }}'
-{{- end }}
-replicaCount: 2
-s3:
-  bucket: registry
-  encrypt: false
-  region: us-east-1
-  regionEndpoint: DYNAMIC
-  rootdirectory: /registry
-  secure: false
-secrets:
-  s3:
-    secretRef: seaweedfs-s3-rw
-storage: s3
-service:
-  name: registry
-  type: ClusterIP
-  port: 5000
