--- pkg/addons/registry/static/values.tpl.yaml
@@ -1,31 +1,0 @@
-configData:
-  auth:
-    htpasswd:
-      path: /auth/htpasswd
-      realm: Registry
-extraVolumeMounts:
-- mountPath: /auth
-  name: auth
-extraVolumes:
-- name: auth
-  secret:
-    secretName: registry-auth
-fullnameOverride: registry
-{{- if .ReplaceImages }}
-image:
-  repository: '{{ (index .Images "registry").Repo }}'
-  tag: '{{ index (index .Images "registry").Tag .GOARCH }}'
-{{- end }}
-persistence:
-  accessMode: ReadWriteOnce
-  enabled: true
-  size: 10Gi
-  storageClass: openebs-hostpath
-podAnnotations:
-  backup.velero.io/backup-volumes: data
-replicaCount: 1
-storage: filesystem
-service:
-  name: registry
-  type: ClusterIP
-  port: 5000
