--- pkg/addons/seaweedfs/seaweedfs_test.go
@@ -1,12 +1,0 @@
-package seaweedfs
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func Test_init(t *testing.T) {
-	// init will panic if it fails to unmarshal the helm values
-	assert.True(t, true)
-}
