--- pkg/addons/seaweedfs/static/metadata.yaml
@@ -1,15 +1,0 @@
-#
-# this file is automatically generated by buildtools. manual edits are not recommended.
-# to regenerate this file, run the following commands:
-#
-# $ make buildtools
-# $ output/bin/buildtools update addon <addon name>
-#
-version: 4.0.379
-location: oci://proxy.replicated.com/anonymous/registry.replicated.com/ec-charts/seaweedfs
-images:
-    seaweedfs:
-        repo: proxy.replicated.com/anonymous/replicated/ec-seaweedfs
-        tag:
-            amd64: 3.79-r0-amd64@sha256:f86f30cf779f30a08293db302141cb3d2bf693b23d2c4af64575ce00c26cc783
-            arm64: 3.79-r0-arm64@sha256:bfe68bc9e38fd4686a0ba5c825600ccd2dc18343f613f574a240330d34c6eb42
