--- pkg/addons/velero/static/metadata.yaml
@@ -1,30 +1,0 @@
-#
-# this file is automatically generated by buildtools. manual edits are not recommended.
-# to regenerate this file, run the following commands:
-#
-# $ make buildtools
-# $ output/bin/buildtools update addon <addon name>
-#
-version: 8.3.0
-location: oci://proxy.replicated.com/anonymous/registry.replicated.com/ec-charts/velero
-images:
-    kubectl:
-        repo: proxy.replicated.com/anonymous/replicated/ec-kubectl
-        tag:
-            amd64: 1.32.1-r2-amd64@sha256:479452dd00b3d2b17b5e0a93ca6f376a781608e906238685376ab233a6f1af7c
-            arm64: 1.32.1-r2-arm64@sha256:df0d96453f2fe340801f1031419245f97a22d5a4abaa5b7039fc5fd894d34f64
-    velero:
-        repo: proxy.replicated.com/anonymous/replicated/ec-velero
-        tag:
-            amd64: 1.15.2-r1-amd64@sha256:0fb1e230ac0bfe2c812cd38029a00edbd37fb090955194852aef711bc3f7fbf2
-            arm64: 1.15.2-r1-arm64@sha256:2a1fd4add3ab64ff8787aa1c27728819ac68bb1c27741046619c8d668e4215ed
-    velero-plugin-for-aws:
-        repo: proxy.replicated.com/anonymous/replicated/ec-velero-plugin-for-aws
-        tag:
-            amd64: 1.11.1-r1-amd64@sha256:780e1d6e3d02f8f84e13620ea1479e58d84a4a671444d6cdbf8d20aee5f4c945
-            arm64: 1.11.1-r1-arm64@sha256:30d2e59446640bb82e237875232e833dab7eeee6793e9e8b0564fe69d4316896
-    velero-restore-helper:
-        repo: proxy.replicated.com/anonymous/replicated/ec-velero-restore-helper
-        tag:
-            amd64: 1.15.2-r1-amd64@sha256:cd9153dc25d4291fda712b5659bffc78ac45c6bbc07b49862b72ab5c72c13d11
-            arm64: 1.15.2-r1-arm64@sha256:2c6145aad99ff9e1edb24b25dc91a282f426820c05dcfcb9414acf47c38cf75f
