--- pkg/addons/velero/static/values.tpl.yaml
@@ -1,30 +1,0 @@
-backupsEnabled: false
-credentials:
-  existingSecret: cloud-credentials
-deployNodeAgent: true
-{{- if .ReplaceImages }}
-image:
-  repository: '{{ (index .Images "velero").Repo }}'
-  tag: '{{ index (index .Images "velero").Tag .GOARCH }}'
-initContainers:
-- image: '{{ ImageString (index .Images "velero-plugin-for-aws") }}'
-  imagePullPolicy: IfNotPresent
-  name: velero-plugin-for-aws
-  volumeMounts:
-  - mountPath: /target
-    name: plugins
-kubectl:
-  image:
-    repository: '{{ (index .Images "kubectl").Repo }}'
-    tag: '{{ index (index .Images "kubectl").Tag .GOARCH }}'
-configMaps:
-  fs-restore-action-config:
-    labels:
-      velero.io/plugin-config: ""
-      velero.io/pod-volume-restore: RestoreItemAction
-    data:
-      image: '{{ ImageString (index .Images "velero-restore-helper") }}'
-{{- end }}
-nodeAgent:
-  podVolumePath: /var/lib/embedded-cluster/k0s/kubelet/pods
-snapshotsEnabled: false
