+++ tests/integration/kind/openebs/analytics_test.go
@@ -1,9 +1,11 @@
 package openebs
 
 import (
+	"context"
 	"testing"
 
-	"github.com/replicatedhq/embedded-cluster/pkg/addons/openebs"
+	"github.com/replicatedhq/embedded-cluster/pkg/addons2/openebs"
+	"github.com/replicatedhq/embedded-cluster/pkg/helm"
 	"github.com/replicatedhq/embedded-cluster/pkg/runtimeconfig"
 	"github.com/replicatedhq/embedded-cluster/tests/integration/util"
 	"github.com/stretchr/testify/assert"
@@ -15,19 +17,18 @@ func TestOpenEBS_AnalyticsDisabled(t *testing.T) {
 	clusterName := util.GenerateClusterName(t)
 	kubeconfig := util.SetupKindCluster(t, clusterName, nil)
 
-	addon := openebs.OpenEBS{}
 	runtimeconfig.SetDataDir("/custom")
-	charts, _, err := addon.GenerateHelmConfig(nil, false)
-	require.NoError(t, err, "failed to generate helm config")
 
-	chart := charts[0]
-	namespace := chart.TargetNS
+	hcli, err := helm.NewClient(helm.HelmOptions{KubeConfig: kubeconfig})
+	defer hcli.Close()
+	require.NoError(t, err)
 
-	helmValuesFile := util.WriteHelmValuesFile(t, chart.Values)
+	addon := &openebs.OpenEBS{}
+	if err := addon.Install(context.Background(), nil, hcli, nil, nil); err != nil {
+		t.Fatalf("failed to install openebs: %v", err)
+	}
 
-	util.HelmInstall(t, kubeconfig, namespace, chart.Name, chart.Version, chart.ChartName, helmValuesFile)
-
-	deploy := util.GetDeployment(t, kubeconfig, namespace, "openebs-localpv-provisioner")
+	deploy := util.GetDeployment(t, kubeconfig, addon.Namespace(), "openebs-localpv-provisioner")
 	assert.Contains(t, deploy.Spec.Template.Spec.Containers[0].Env,
 		corev1.EnvVar{Name: "OPENEBS_IO_ENABLE_ANALYTICS", Value: "false"},
 		"openebs should not send analytics")
