+++ tests/integration/kind/openebs/customdatadir_test.go
@@ -1,12 +1,14 @@
 package openebs
 
 import (
+	"context"
 	"os"
 	"path/filepath"
 	"testing"
 	"time"
 
-	"github.com/replicatedhq/embedded-cluster/pkg/addons/openebs"
+	"github.com/replicatedhq/embedded-cluster/pkg/addons2/openebs"
+	"github.com/replicatedhq/embedded-cluster/pkg/helm"
 	"github.com/replicatedhq/embedded-cluster/pkg/runtimeconfig"
 	"github.com/replicatedhq/embedded-cluster/tests/integration/util"
 	"github.com/replicatedhq/embedded-cluster/tests/integration/util/kind"
@@ -27,17 +29,16 @@ func TestOpenEBS_CustomDataDir(t *testing.T) {
 	})
 	kubeconfig := util.SetupKindClusterFromConfig(t, kindConfig)
 
-	addon := openebs.OpenEBS{}
 	runtimeconfig.SetDataDir("/custom")
-	charts, _, err := addon.GenerateHelmConfig(nil, false)
-	require.NoError(t, err, "failed to generate helm config")
 
-	chart := charts[0]
-	namespace := chart.TargetNS
+	hcli, err := helm.NewClient(helm.HelmOptions{KubeConfig: kubeconfig})
+	defer hcli.Close()
+	require.NoError(t, err)
 
-	helmValuesFile := util.WriteHelmValuesFile(t, chart.Values)
-
-	util.HelmInstall(t, kubeconfig, namespace, chart.Name, chart.Version, chart.ChartName, helmValuesFile)
+	addon := &openebs.OpenEBS{}
+	if err := addon.Install(context.Background(), nil, hcli, nil, nil); err != nil {
+		t.Fatalf("failed to install openebs: %v", err)
+	}
 
 	util.WaitForStorageClass(t, kubeconfig, "openebs-hostpath", 30*time.Second)
